Commit 3
Troca de elementos das matrizes de inteiro para real
Formatação na saída de dados dos elementos das matrizes
Adição de comentários
Atualização no README.md

+++ src/Calculos.java
@@ -1,12 +1,12 @@
 public class Calculos {
-    public int[][] somarMatrizes(Matriz matriz1, Matriz matriz2) {
+    public float[][] somarMatrizes(Matriz matriz1, Matriz matriz2) {
         if (matriz1.getLinhas() != matriz2.getLinhas()
                 || matriz2.getColunas() != matriz2.getColunas()) {
             System.out.println("As matrizes precisam ser iguais em números de linhas e colunas.\n");
             return null;
         }
 
-        int[][] matrizResultante = new int[matriz1.getLinhas()][matriz2.getColunas()];
+        float[][] matrizResultante = new float[matriz1.getLinhas()][matriz2.getColunas()];
         for (int i = 0; i < matrizResultante.length; i++){
             for (int j = 0; j < matrizResultante[0].length; j++){
                 matrizResultante[i][j] = matriz1.getElementos()[i][j] + matriz2.getElementos()[i][j];
@@ -16,14 +16,14 @@ public class Calculos {
         return matrizResultante;
     }
 
-    public int[][] subtrairMatrizes(Matriz matriz1, Matriz matriz2) {
+    public float[][] subtrairMatrizes(Matriz matriz1, Matriz matriz2) {
         if (matriz1.getLinhas() != matriz2.getLinhas()
                 || matriz2.getColunas() != matriz2.getColunas()) {
             System.out.println("As matrizes precisam ser iguais em números de linhas e colunas.\n");
             return null;
         }
 
-        int[][] matrizResultante = new int[matriz1.getLinhas()][matriz2.getColunas()];
+        float[][] matrizResultante = new float[matriz1.getLinhas()][matriz2.getColunas()];
         for (int i = 0; i < matrizResultante.length; i++){
             for (int j = 0; j < matrizResultante[0].length; j++){
                 matrizResultante[i][j] = matriz1.getElementos()[i][j] - matriz2.getElementos()[i][j];
@@ -33,21 +33,21 @@ public class Calculos {
         return matrizResultante;
     }
 
-    public int[][] multiplicarMatrizes(Matriz matriz1, Matriz matriz2) {
+    public float[][] multiplicarMatrizes(Matriz matriz1, Matriz matriz2) {
         if (matriz1.getElementos()[0].length != matriz2.getElementos().length) {
             System.out.println(
                     "O número de colunas da primeira matriz deve ser igual ao número de linhas da segunda matriz.\n");
             return null;
         }
 
-        int[][] matrizResultante = new int[matriz1.getElementos().length][matriz2.getElementos()[0].length];
+        float[][] matrizResultante = new float[matriz1.getElementos().length][matriz2.getElementos()[0].length];
         for (int i = 0; i < matrizResultante.length; i++) {
             for (int j = 0; j < matrizResultante[0].length; j++) {
 
                 //Calcular elemento [i][j] da matriz resultante
-                int somatorio = 0;
+                float somatorio = 0;
                 for (int aux = 0; aux < matriz1.getElementos()[0].length; aux++) {
-                    int mult = matriz1.getElementos()[i][aux] * matriz2.getElementos()[aux][j];
+                    float mult = matriz1.getElementos()[i][aux] * matriz2.getElementos()[aux][j];
                     somatorio += mult;
                 }
                 matrizResultante[i][j] = somatorio;
